% Auto-generated: do not edit by hand
\name{''Configurator}

\alias{''Configurator}

\title{Configurator component}

\description{
Configurator component
}

\usage{
''Configurator(id=NULL, config=NULL, meta=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}

\item{config}{Prop The resulting configuration of the plot.}

\item{meta}{The metadata the plotter selection is based on.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

