% Auto-generated: do not edit by hand
\name{''MeltTransform}

\alias{''MeltTransform}

\title{MeltTransform component}

\description{

}

\usage{
''MeltTransform(allColOptions=NULL, catColOptions=NULL, config=NULL,
meta=NULL, numColOptions=NULL)
}

\arguments{
\item{allColOptions}{All currently available column options}

\item{catColOptions}{Currently available categorical options}

\item{config}{The config the user sets in this component.}

\item{meta}{The metadata this section is based on.}

\item{numColOptions}{Currently available numerical options}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

