% Auto-generated: do not edit by hand
\name{''PlotterBase}

\alias{''PlotterBase}

\title{PlotterBase component}

\description{

}

\usage{
''PlotterBase(id=NULL, allColOptions=NULL, catColOptions=NULL,
config=NULL, numColOptions=NULL, numOptions=NULL)
}

\arguments{
\item{id}{The config the user sets in this component.}

\item{allColOptions}{All currently available column options}

\item{catColOptions}{Currently available categorical options}

\item{config}{The config the user sets in this component.}

\item{numColOptions}{Currently available numerical options}

\item{numOptions}{Currently available options without grouping}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

