% Auto-generated: do not edit by hand
\name{base}

\alias{base}

\title{Base component}

\description{

}

\usage{
base(id=NULL, config=NULL, meta=NULL, meta_out=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}

\item{config}{The config the user sets in this component.}

\item{meta}{The metadata this section is based on.}

\item{meta_out}{The metadata section will create as output.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

