% Auto-generated: do not edit by hand
\name{table}

\alias{table}

\title{Table component}

\description{

}

\usage{
table(allColOptions=NULL, catColOptions=NULL, config=NULL,
numColOptions=NULL)
}

\arguments{
\item{allColOptions}{All currently available column options}

\item{catColOptions}{Currently available categorical options}

\item{config}{The config the user sets in this component.}

\item{numColOptions}{Currently available numerical options}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

