% Auto-generated: do not edit by hand
\name{''Configurator}

\alias{''Configurator}

\title{Configurator component}

\description{
Configurator component
}

\usage{
''Configurator(id=NULL, config=NULL, meta=NULL, showFilter=NULL,
showMetadata=NULL, showParameterization=NULL,
showPlotter=NULL, showStore=NULL, showTransform=NULL)
}

\arguments{
\item{id}{The ID used to identify this component in Dash callbacks.}

\item{config}{Prop The resulting configuration of the plot.}

\item{meta}{The metadata the plotter selection is based on.}

\item{showFilter}{Prop to define the visibility of the Filter panel}

\item{showMetadata}{Prop to define the visibility of the Metadata panel}

\item{showParameterization}{Prop to define the visibility of the Parameterization panel}

\item{showPlotter}{Prop to define the visibility of the Plot panel}

\item{showStore}{Prop to define the visibility of the Store panel}

\item{showTransform}{Prop to define the visibility of the Transform panel}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

